requirements.txt
python-dotenv==1.0.0
google-generativeai==0.3.2
streamlit==1.28.0
openai==1.3.0